Refactor HTMLWatcher into generic DOMWatcher + HTML impl
